Remove cyclic dependency between Student and Room using Allocation association class and fix Delete and Edit student bug

+++ src/main/java/seedu/resireg/logic/commands/DeleteCommand.java
@@ -8,6 +8,7 @@ import seedu.resireg.commons.core.Messages;
 import seedu.resireg.commons.core.index.Index;
 import seedu.resireg.logic.commands.exceptions.CommandException;
 import seedu.resireg.model.Model;
+import seedu.resireg.model.allocation.Allocation;
 import seedu.resireg.model.student.Student;
 
 /**
@@ -18,6 +19,8 @@ public class DeleteCommand extends Command {
     public static final String COMMAND_WORD = "delete";
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Student: %1$s";
+    public static final String MESSAGE_ROOM_ALLOCATION_EXISTS =
+            "Student %1$s currently has a room! Deallocate student first!";
 
     public static final Help HELP = new Help(COMMAND_WORD,
             "Deletes the student identified by the index number used in the displayed student list.",
@@ -33,12 +36,18 @@ public class DeleteCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
         List<Student> lastShownList = model.getFilteredStudentList();
+        List<Allocation> lastShownAllocationList = model.getFilteredAllocationList();
 
         if (targetIndex.getZeroBased() >= lastShownList.size()) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
         Student studentToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        if (model.isAllocated(studentToDelete)) {
+            throw new CommandException(String.format(MESSAGE_ROOM_ALLOCATION_EXISTS, studentToDelete));
+        }
+
         model.deleteStudent(studentToDelete);
         model.saveStateResiReg();
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, studentToDelete));
